refactor(web): split domain/students.ts into operation sub-modules behind a barrel
Decompose the 2812-line students domain monolith into cohesive ./students/*
sub-modules (enrollment, bulkEnrollment, rosterSync, roleWrites, rosterMigration,
inviteRoster, unenroll, updateStudent) over a leaf rosterPrimitives.ts holding the
dense shared cluster (rosterWriteTree, the resolveClassroomTeam* resolvers,
tryAddUserToTeam, resolveClassroomPendingInvite, matchesRosterRow, retryDeferred,
listClassroomMembersWithRoles, the error classes, scoped log). Every operation
module imports these downward from the leaf; withGitConflictRetry stays in
domain/classrooms.ts (shared with assignments).

students.ts is now a thin re-export barrel: all direct exports preserved plus the
two @/util/rosterCsv re-export blocks kept verbatim, so all ~27 importers are
unchanged. Behavior-preserving — tsc + full vitest (1533) green, no-cycle +
boundaries + arch:validate green; students.test.ts still exercises the public
surface via the barrel.
